Generate model using dataset and webcam support
`train_model.py` has been added to train the model using the
dataset from CK+ database. Scripts have been added to prepare this
dataset so that it can be used to generate the model.

Added README

A trained model has been added which can be used to read emotions
from frames. This is a linear SVM model.