Improve prediction by resizing images
A frame is now resized into a new frame for each face, with the
same dimensions as the images used to train the model (350x350)
This causes the faces to be properly predicted by the model.

The real-time webcam predictor now adds the emotion image to the face
of each person it detects.

Two changes have been made to better train the model:
    - Happiness was never trained, because in the dataset, the folder
    is called 'happy', rather than 'happiness'. happiness renamed
    to happy to fix this.
    - The googleset from Paul van Gent's article has been added to
    the data/ folder and is now also used to train the model.

A couple of stability issues still remain. Resizing images seems to
fail from time to time. Although not yet tested, this might have to
do with tilted faces. Putting the text with the prediction on the
frame also seems to crash the program sometimes, with an out of bounds
like error.