Bind most functionality under one interface
A command line runner has been created to allow for accessing all
the different kinds of functionality through one common interface.
So far, sorting of the CK+ dataset and predicting emotions from
the webcam have been added.

Sorting has been reworked into cleaner, usable code.

There is a now a record function that is able to record a video
with a filename, dimensions, duration, fps and codec. This is,
however, a blocking function, so it should be started in a
separate process to ensure other code execution is not halted.